add types folder for calendar and workshop type
