+++ src/types/calendar.ts
@@ -1,0 +1,6 @@
+import type {workshop} from '../types/workshop.ts';
+
+export type calendar = {
+  day: string;
+  workshops: workshop[];
+}
